Seperated handlers & logging fixes
